+++ ElvUI/Wrath/Modules/Skins/Character.lua
@@ -7,16 +7,15 @@ local unpack = unpack
 local ipairs, pairs = ipairs, pairs
 
 local HasPetUI = HasPetUI
+local hooksecurefunc = hooksecurefunc
 local GetPetHappiness = GetPetHappiness
 local GetSkillLineInfo = GetSkillLineInfo
-local GetInventoryItemQuality = GetInventoryItemQuality
-local GetItemQualityColor = GetItemQualityColor
 local UnitFactionGroup = UnitFactionGroup
-local hooksecurefunc = hooksecurefunc
+local GetItemQualityColor = GetItemQualityColor
+local GetInventoryItemQuality = GetInventoryItemQuality
 
-local MAX_ARENA_TEAMS = MAX_ARENA_TEAMS
-local NUM_COMPANIONS_PER_PAGE = NUM_COMPANIONS_PER_PAGE
 local NUM_FACTIONS_DISPLAYED = NUM_FACTIONS_DISPLAYED
+local NUM_COMPANIONS_PER_PAGE = NUM_COMPANIONS_PER_PAGE
 local CHARACTERFRAME_SUBFRAMES = CHARACTERFRAME_SUBFRAMES
 
 local HONOR_CURRENCY = Constants.CurrencyConsts.CLASSIC_HONOR_CURRENCY_ID
@@ -29,11 +28,23 @@ local ResistanceCoords = {
 	{ 0.21875, 0.8125, 0.4765625, 0.55078125},	--Shadow
 }
 
-local petClasses = {
-	HUNTER = true,
-	WARLOCK = true,
-	DEATHKNIGHT = true
-}
+local function HandleTabs()
+	_G.PetPaperDollFrameTab1:ClearAllPoints()
+	_G.PetPaperDollFrameTab1:Point('TOPLEFT', _G.PetPaperDollFrameCompanionFrame, 88, -40)
+
+	local lastTab
+	for index, tab in next, { _G.CharacterFrameTab1, HasPetUI() and _G.CharacterFrameTab2 or nil, _G.CharacterFrameTab3, _G.CharacterFrameTab4, _G.CharacterFrameTab5 } do
+		tab:ClearAllPoints()
+
+		if index == 1 then
+			tab:Point('TOPLEFT', _G.CharacterFrame, 'BOTTOMLEFT', 1, 76)
+		else
+			tab:Point('TOPLEFT', lastTab, 'TOPRIGHT', -19, 0)
+		end
+
+		lastTab = tab
+	end
+end
 
 local function Update_GearManagerDialogPopup()
 	_G.GearManagerDialogPopup:ClearAllPoints()
@@ -183,18 +194,8 @@ function S:CharacterFrame()
 	end
 
 	-- Reposition Tabs
-	_G.CharacterFrameTab1:ClearAllPoints()
-	_G.CharacterFrameTab1:Point('TOPLEFT', _G.CharacterFrame, 'BOTTOMLEFT', 1, 76)
-
-	if petClasses[E.myclass] then
-		_G.CharacterFrameTab2:Point('TOPLEFT', _G.CharacterFrameTab1, 'TOPRIGHT', -19, 0)
-		_G.CharacterFrameTab3:Point('TOPLEFT', _G.CharacterFrameTab2, 'TOPRIGHT', -19, 0)
-	else
-		_G.CharacterFrameTab3:Point('TOPLEFT', _G.CharacterFrameTab1, 'TOPRIGHT', -19, 0)
-	end
-
-	_G.CharacterFrameTab4:Point('TOPLEFT', _G.CharacterFrameTab3, 'TOPRIGHT', -19, 0)
-	_G.CharacterFrameTab5:Point('TOPLEFT', _G.CharacterFrameTab4, 'TOPRIGHT', -19, 0)
+	hooksecurefunc('PetPaperDollFrame_UpdateTabs', HandleTabs)
+	HandleTabs()
 
 	-- HandleTab looks weird
 	for i = 1, 3 do
@@ -203,11 +204,6 @@ function S:CharacterFrame()
 		tab:Height(24)
 		S:HandleButton(tab)
 	end
-
-	hooksecurefunc('PetPaperDollFrame_UpdateTabs', function()
-		_G.PetPaperDollFrameTab1:ClearAllPoints()
-		_G.PetPaperDollFrameTab1:Point('TOPLEFT', _G.PetPaperDollFrameCompanionFrame, 'TOPLEFT', 88, -40)
-	end)
 
 	_G.PaperDollFrame:StripTextures()
 
@@ -499,61 +495,6 @@ function S:CharacterFrame()
 
 	S:HandleCloseButton(_G.SkillDetailStatusBarUnlearnButton)
 	_G.SkillDetailStatusBarUnlearnButton:Point('LEFT', _G.SkillDetailStatusBarBorder, 'RIGHT', -6, 1)
-
-	-- Honor/Arena/PvP Tab
-	local PVPFrame = _G.PVPFrame
-	S:HandleFrame(PVPFrame, true, nil, 11, -12, -32, 76)
-	S:HandleCloseButton(_G.PVPParentFrameCloseButton)
-	_G.PVPParentFrameCloseButton:Point('TOPRIGHT', -26, -5)
-
-	for i = 1, MAX_ARENA_TEAMS do
-		local pvpTeam = _G['PVPTeam'..i]
-
-		pvpTeam:StripTextures()
-		pvpTeam:CreateBackdrop()
-		pvpTeam.backdrop:Point('TOPLEFT', 9, -4)
-		pvpTeam.backdrop:Point('BOTTOMRIGHT', -24, 3)
-
-		pvpTeam:HookScript('OnEnter', S.SetModifiedBackdrop)
-		pvpTeam:HookScript('OnLeave', S.SetOriginalBackdrop)
-
-		_G['PVPTeam'..i..'Highlight']:Kill()
-	end
-
-	local PVPTeamDetails = _G.PVPTeamDetails
-	PVPTeamDetails:StripTextures()
-	PVPTeamDetails:SetTemplate('Transparent')
-	PVPTeamDetails:Point('TOPLEFT', PVPFrame, 'TOPRIGHT', -30, -12)
-
-	local PVPFrameToggleButton = _G.PVPFrameToggleButton
-	S:HandleNextPrevButton(PVPFrameToggleButton)
-	PVPFrameToggleButton:Point('BOTTOMRIGHT', PVPFrame, 'BOTTOMRIGHT', -48, 81)
-	PVPFrameToggleButton:Size(14)
-
-	for i = 1, 5 do
-		local header = _G['PVPTeamDetailsFrameColumnHeader'..i]
-		header:StripTextures()
-		header:StyleButton()
-	end
-
-	for i = 1, 10 do
-		local button = _G['PVPTeamDetailsButton'..i]
-		button:Width(335)
-		S:HandleButtonHighlight(button)
-	end
-
-	-- BG Queue Tabs
-	S:HandleTab(_G.PVPParentFrameTab1)
-	S:HandleTab(_G.PVPParentFrameTab2)
-
-	-- Reposition Tabs
-	_G.PVPParentFrameTab1:ClearAllPoints()
-	_G.PVPParentFrameTab1:Point('TOPLEFT', _G.PVPParentFrame, 'BOTTOMLEFT', 1, 76)
-	_G.PVPParentFrameTab2:Point('TOPLEFT', _G.PVPParentFrameTab1, 'TOPRIGHT', -19, 0)
-
-	S:HandleButton(_G.PVPTeamDetailsAddTeamMember)
-	S:HandleNextPrevButton(_G.PVPTeamDetailsToggleButton)
-	S:HandleCloseButton(_G.PVPTeamDetailsCloseButton)
 
 	-- TokenFrame (Currency Tab)
 	_G.TokenFrame:StripTextures()