+++ ElvUI_Options/Core/Skins.lua
@@ -59,10 +59,7 @@ if E.Wrath or E.Retail then
 	toggles.collections = L["COLLECTIONS"]
 	toggles.gbank = L["Guild Bank"]
 	toggles.pvp = L["PvP Frames"]
-end
-
-if E.Wrath or E.Classic then
-	toggles.battlefield = L["Battlefield"]
+	toggles.lfg = L["LFG_TITLE"]
 end
 
 if E.Retail then
@@ -114,8 +111,8 @@ if E.Retail then
 elseif E.Wrath then
 	toggles.arena = L["Arena"]
 	toggles.arenaRegistrar = L["Arena Registrar"]
-	toggles.lfg = L["LFG_TITLE"]
 elseif E.Classic then
+	toggles.battlefield = L["Battlefield"]
 	toggles.craft = L["Craft"]
 end